Removal of UserDefinedProtocol.h
the file UserDefinedProtocol.h has been removed and all dependancies on
it have been changed. code builds.
lost functionality to correctly read userdefined protocols from file,
they are still added to the database, just not given to the ui correctly

many functions added to be able to modify protocols in the database.
this is nessicary to beable to change an added protocol.
minimum req to adding this branch to develop is to have all the
functionality that was there. so, that is where we are headed.