changes to network protocol db XML file
In preperation of removing the lists from networks for protocols
i am redoing all networks that have more than one detail in a in a
single network's src or des.
Basicly the ability to have 1 source and many destinations, or many
sources and 1 destination, or many sources and many destinations, is
being changed. Before it was possible to define a net that operates with
src 1, 4, 6 and des 3, 6. this would expand into
1 to 3, 1 to 6, 4 to 3, 4 to 6, 6 to 3, 6 to 6.
Now it seems like having that ability would be great, and really it
could be, IF IT WAS EVER USED. Unfortunatly, the complexity of having a
vector<protocolentry> where every protocolentry has a vector<network>
and each network has a vector<detail> for each src and des isn't
worth having that functionality, especially when it's functionally
equivilant to expanding the networks by hand, and would only ever help
someone adding very strange entries to the XML file. it makes things
more complicated everywhere... and was only used for about 6 protocols.
So there is my reasoning behind this change.

I plan on ripping out the vectors<detail> and refactoring all the code
that uses them to just look at a specific detail for each.
There is one other little problem where some protocols have no src or
des. This _only_ ever happens if it is icmp, because i guess iptables
has some special things it does for them. In these cases a check of size
will no longer work to see if it is empty, checking the protocol type
before using the detail will basicly be required for safe operation.