removal of Ranges of UDPs added
The ability to remove a range is now avalible.
There is a check to ensure that the last connection
is not removed. A UDP must have at least one
port range, (this isn't the case of all protocols because some use the
blank range to mean everything allowed, but because this
is kindof counter intuitive a user shouldn't be allowed to accidentally
do it)