regression fix
Changed the entries for ftp and ssh back to how they were originally,
so that conections from guarddog builds still work with this version.

There is almost no way to know which udp is which number anymore,
support for that has been removed. Changed the error for udp import
to let the user know that the zone connections were not able to be
imported for the udp. That is the best we can do.