user experiance issues cleared up.
The interface for manipulating userdefined protocols is now more
responsive and does things more inline with proper "expected" behaviour,
with new user defined protocols being auto opened. currently on deletion
of a protocol the last one is oppened but i think i can change that to
be the one above the previous.