when adding new udps the names will be unique
this fixes potential problems where the name and indexes may not
nessicarily be the same. like if there are 3 "test" protocols
adding sub ranges to the third will work, but modifiying it will modify
the first, causing memory coruption.
this is because the lookup function returns the first matching protocol,
not neccicarily the correct one.